[ApiBundle] Update QueryItemExtension & QueryExtension classes
