+++ src/Sylius/Bundle/ApiBundle/Doctrine/QueryItemExtension/EnabledProductVariantItemExtension.php
@@ -13,14 +13,15 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ApiBundle\Doctrine\QueryItemExtension;
 
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Extension\QueryItemExtensionInterface;
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Util\QueryNameGeneratorInterface;
+use ApiPlatform\Doctrine\Orm\Extension\QueryItemExtensionInterface;
+use ApiPlatform\Doctrine\Orm\Util\QueryNameGeneratorInterface;
+use ApiPlatform\Metadata\Operation;
 use Doctrine\ORM\QueryBuilder;
 use Sylius\Bundle\ApiBundle\Context\UserContextInterface;
 use Sylius\Component\Core\Model\ProductVariantInterface;
 
 /** @experimental */
-final class EnabledProductVariantItemExtension implements QueryItemExtensionInterface
+final readonly class EnabledProductVariantItemExtension implements QueryItemExtensionInterface
 {
     public function __construct(private UserContextInterface $userContext)
     {
@@ -31,9 +32,9 @@ final class EnabledProductVariantItemExtension implements QueryItemExtensionInte
         QueryNameGeneratorInterface $queryNameGenerator,
         string $resourceClass,
         array $identifiers,
-        string $operationName = null,
+        Operation $operation = null,
         array $context = [],
-    ) {
+    ): void {
         if (!is_a($resourceClass, ProductVariantInterface::class, true)) {
             return;
         }
