+++ src/Sylius/Bundle/ApiBundle/spec/Doctrine/QueryItemExtension/OrderVisitorItemExtensionSpec.php
@@ -14,53 +14,97 @@ declare(strict_types=1);
 namespace spec\Sylius\Bundle\ApiBundle\Doctrine\QueryItemExtension;
 
 use ApiPlatform\Doctrine\Orm\Util\QueryNameGeneratorInterface;
+use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\Get;
+use ApiPlatform\Metadata\Put;
+use Doctrine\ORM\Query\Expr;
 use Doctrine\ORM\QueryBuilder;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Sylius\Bundle\ApiBundle\Context\UserContextInterface;
 use Sylius\Bundle\ApiBundle\Serializer\ContextKeys;
-use Sylius\Component\Core\Model\CustomerInterface;
 use Sylius\Component\Core\Model\OrderInterface;
-use Sylius\Component\Core\Model\ShopUserInterface;
 use Sylius\Component\User\Model\UserInterface;
 use Symfony\Component\HttpFoundation\Request;
 
-final class OrderShopUserItemExtensionSpec extends ObjectBehavior
+final class OrderVisitorItemExtensionSpec extends ObjectBehavior
 {
     function let(UserContextInterface $userContext): void
     {
-        $this->beConstructedWith($userContext, ['shop_select_payment_method', 'shop_account_change_payment_method']);
+        $this->beConstructedWith($userContext, ['shop_select_payment_method']);
     }
 
-    function it_filters_carts_for_shop_users_to_the_one_owned_by_them_for_methods_other_than_get(
+    function it_filters_carts_for_visitors_to_not_authorized_for_methods_other_than_get(
         UserContextInterface $userContext,
         QueryBuilder $queryBuilder,
-        ShopUserInterface $shopUser,
-        CustomerInterface $customer,
         QueryNameGeneratorInterface $queryNameGenerator,
+        Expr $expr,
     ): void {
+        $queryNameGenerator->generateParameterName('state')->shouldBeCalled()->willReturn('state');
         $queryBuilder->getRootAliases()->willReturn(['o']);
 
-        $userContext->getUser()->willReturn($shopUser);
+        $userContext->getUser()->willReturn(null);
 
-        $shopUser->getCustomer()->willReturn($customer);
-        $customer->getId()->willReturn(1);
+        $queryBuilder
+            ->leftJoin(sprintf('%s.customer', 'o'), 'customer')
+            ->shouldBeCalled()
+            ->willReturn($queryBuilder)
+        ;
 
-        $queryNameGenerator->generateParameterName('state')->shouldBeCalled()->willReturn('state');
-        $queryNameGenerator->generateParameterName('customer')->shouldBeCalled()->willReturn('customer');
+        $queryBuilder
+            ->leftJoin('customer.user', 'user')
+            ->shouldBeCalled()
+            ->willReturn($queryBuilder)
+        ;
 
-        $queryBuilder->andWhere('o.customer = :customer')->shouldBeCalled()->willReturn($queryBuilder);
-        $queryBuilder->setParameter('customer', 1)->shouldBeCalled()->willReturn($queryBuilder);
+        $queryBuilder
+            ->expr()
+            ->shouldBeCalled()
+            ->willReturn($expr)
+        ;
 
-        $queryBuilder->andWhere('o.state = :state')->shouldBeCalled()->willReturn($queryBuilder);
-        $queryBuilder->setParameter('state', OrderInterface::STATE_CART)->shouldBeCalled()->willReturn($queryBuilder);
+        $queryBuilder
+            ->setParameter('createdByGuest', true)
+            ->shouldBeCalled()
+            ->willReturn($expr)
+        ;
+
+        $expr
+            ->andX('o.customer IS NOT NULL', 'o.createdByGuest = :createdByGuest')
+            ->shouldBeCalled()
+            ->willReturn('o.customer IS NOT NULL AND o.createdByGuest = :createdByGuest')
+        ;
+
+        $expr
+            ->orX('user IS NULL', 'o.customer IS NULL', 'o.customer IS NOT NULL AND o.createdByGuest = :createdByGuest')
+            ->shouldBeCalled()
+            ->willReturn('user IS NULL OR o.customer IS NULL OR (o.customer IS NOT NULL AND o.createdByGuest = :createdByGuest)')
+        ;
+
+        $queryBuilder
+            ->andWhere('user IS NULL OR o.customer IS NULL OR (o.customer IS NOT NULL AND o.createdByGuest = :createdByGuest)')
+            ->shouldBeCalled()
+            ->willReturn($queryBuilder)
+        ;
+
+        $queryBuilder
+            ->andWhere(sprintf('%s.state = :state', 'o'))
+            ->shouldBeCalled()
+            ->willReturn($queryBuilder)
+        ;
+
+        $queryBuilder
+            ->setParameter('state', OrderInterface::STATE_CART)
+            ->shouldBeCalled()
+            ->willReturn($queryBuilder)
+        ;
 
         $this->applyToItem(
             $queryBuilder,
             $queryNameGenerator,
             OrderInterface::class,
             ['tokenValue' => 'xaza-tt_fee'],
-            Request::METHOD_PUT,
+            new Delete(),
             [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_POST],
         );
 
@@ -69,7 +113,7 @@ final class OrderShopUserItemExtensionSpec extends ObjectBehavior
             $queryNameGenerator,
             OrderInterface::class,
             ['tokenValue' => 'xaza-tt_fee'],
-            Request::METHOD_PUT,
+            new Delete(),
             [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_PATCH],
         );
 
@@ -78,7 +122,7 @@ final class OrderShopUserItemExtensionSpec extends ObjectBehavior
             $queryNameGenerator,
             OrderInterface::class,
             ['tokenValue' => 'xaza-tt_fee'],
-            Request::METHOD_PUT,
+            new Delete(),
             [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_PUT],
         );
 
@@ -87,36 +131,69 @@ final class OrderShopUserItemExtensionSpec extends ObjectBehavior
             $queryNameGenerator,
             OrderInterface::class,
             ['tokenValue' => 'xaza-tt_fee'],
-            Request::METHOD_PUT,
+            new Delete(),
             [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_DELETE],
         );
     }
 
-    function it_filters_carts_and_orders_for_shop_users_to_the_one_owned_by_them_for_get_and_payment_selection_operations(
+    function it_filters_orders_for_visitors_to_not_authorized_orders_for_get_operations_and_payment_selection(
         UserContextInterface $userContext,
         QueryBuilder $queryBuilder,
-        ShopUserInterface $shopUser,
-        CustomerInterface $customer,
         QueryNameGeneratorInterface $queryNameGenerator,
+        Expr $expr,
     ): void {
         $queryBuilder->getRootAliases()->willReturn(['o']);
 
-        $userContext->getUser()->willReturn($shopUser);
+        $userContext->getUser()->willReturn(null);
 
-        $shopUser->getCustomer()->willReturn($customer);
-        $customer->getId()->willReturn(1);
+        $queryBuilder
+            ->leftJoin(sprintf('%s.customer', 'o'), 'customer')
+            ->shouldBeCalled()
+            ->willReturn($queryBuilder)
+        ;
 
-        $queryNameGenerator->generateParameterName('customer')->shouldBeCalled()->willReturn('customer');
+        $queryBuilder
+            ->leftJoin('customer.user', 'user')
+            ->shouldBeCalled()
+            ->willReturn($queryBuilder)
+        ;
 
-        $queryBuilder->andWhere('o.customer = :customer')->shouldBeCalled()->willReturn($queryBuilder);
-        $queryBuilder->setParameter('customer', 1)->shouldBeCalled()->willReturn($queryBuilder);
+        $queryBuilder
+            ->expr()
+            ->shouldBeCalled()
+            ->willReturn($expr)
+        ;
+
+        $queryBuilder
+            ->setParameter('createdByGuest', true)
+            ->shouldBeCalled()
+            ->willReturn($expr)
+        ;
+
+        $expr
+            ->andX('o.customer IS NOT NULL', 'o.createdByGuest = :createdByGuest')
+            ->shouldBeCalled()
+            ->willReturn('o.customer IS NOT NULL AND o.createdByGuest = :createdByGuest')
+        ;
+
+        $expr
+            ->orX('user IS NULL', 'o.customer IS NULL', 'o.customer IS NOT NULL AND o.createdByGuest = :createdByGuest')
+            ->shouldBeCalled()
+            ->willReturn('user IS NULL OR o.customer IS NULL OR (o.customer IS NOT NULL AND o.createdByGuest = :createdByGuest)')
+        ;
+
+        $queryBuilder
+            ->andWhere('user IS NULL OR o.customer IS NULL OR (o.customer IS NOT NULL AND o.createdByGuest = :createdByGuest)')
+            ->shouldBeCalled()
+            ->willReturn($queryBuilder)
+        ;
 
         $this->applyToItem(
             $queryBuilder,
             $queryNameGenerator,
             OrderInterface::class,
             ['tokenValue' => 'xaza-tt_fee'],
-            Request::METHOD_PUT,
+            new Get(),
             [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_GET],
         );
 
@@ -125,21 +202,12 @@ final class OrderShopUserItemExtensionSpec extends ObjectBehavior
             $queryNameGenerator,
             OrderInterface::class,
             ['tokenValue' => 'xaza-tt_fee'],
-            'shop_select_payment_method',
-            [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_PATCH],
-        );
-
-        $this->applyToItem(
-            $queryBuilder,
-            $queryNameGenerator,
-            OrderInterface::class,
-            ['tokenValue' => 'xaza-tt_fee'],
-            'shop_account_change_payment_method',
+            new Put(name: 'shop_select_payment_method'),
             [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_PATCH],
         );
     }
 
-    function it_does_nothing_if_logged_in_user_is_not_shop_user(
+    function it_does_nothing_if_any_user_is_logged_in(
         UserContextInterface $userContext,
         QueryBuilder $queryBuilder,
         UserInterface $user,
@@ -147,7 +215,7 @@ final class OrderShopUserItemExtensionSpec extends ObjectBehavior
     ): void {
         $queryBuilder->getRootAliases()->willReturn(['o']);
 
-        $userContext->getUser()->willReturn(null, $user);
+        $userContext->getUser()->willReturn($user);
 
         $queryBuilder->leftJoin(Argument::any())->shouldNotBeCalled();
         $queryBuilder->expr()->shouldNotBeCalled();
@@ -159,16 +227,7 @@ final class OrderShopUserItemExtensionSpec extends ObjectBehavior
             $queryNameGenerator,
             OrderInterface::class,
             ['tokenValue' => 'xaza-tt_fee'],
-            Request::METHOD_PUT,
-            [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_PUT],
-        );
-
-        $this->applyToItem(
-            $queryBuilder,
-            $queryNameGenerator,
-            OrderInterface::class,
-            ['tokenValue' => 'xaza-tt_fee'],
-            Request::METHOD_PUT,
+            new Put(),
             [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_PUT],
         );
     }
@@ -187,7 +246,7 @@ final class OrderShopUserItemExtensionSpec extends ObjectBehavior
             $queryNameGenerator,
             \stdClass::class,
             ['tokenValue' => 'xaza-tt_fee'],
-            Request::METHOD_PUT,
+            new Put(),
             [ContextKeys::HTTP_REQUEST_METHOD_TYPE => Request::METHOD_PUT],
         );
     }
